requirements.txt
fastapi
uvicorn
httpx
pyjwt
scikit-learn
pandas
streamlit
python-multipart
